+++ powerhub/repos.py
@@ -5,11 +5,12 @@ from urllib.parse import urlparse
 from powerhub.directories import MOD_DIR
 
 repositories = {
-    "AdrianVollmer/PowerSploit":
-        "https://github.com/AdrianVollmer/PowerSploit.git",
-    "BloodHound": "https://github.com/BloodHoundAD/BloodHound.git",
+    "ZeroDayLab/PowerSploit": "https://github.com/ZeroDayLab/PowerSploit.git",
+    "SharpHound": "https://github.com/BloodHoundAD/BloodHound/raw/master/Collectors/SharpHound.exe",  # noqa
     "ASREPRoast": "https://github.com/HarmJ0y/ASREPRoast.git",
     "Nishang": "https://github.com/samratashok/nishang.git",
+    "PowerSharpPack": "https://github.com/S3cur3Th1sSh1t/PowerSharpPack",
+    "Ghostpack Binaries": "https://github.com/r3motecontrol/Ghostpack-CompiledBinaries",  # noqa
 }
 
 
@@ -37,7 +38,7 @@ def git_clone(url):
     """Installs a git repository"""
     parsed_url = urlparse(url)
     basename = os.path.basename(parsed_url.path)
-    dest_dir = os.path.join(MOD_DIR, 'ps1', basename[:-4])
+    dest_dir = os.path.join(MOD_DIR, basename[:-4])
     if os.path.isdir(dest_dir):
         raise Exception("Directory already exists: %s" % dest_dir)
     subprocess.check_output(['git', 'clone', '--depth', '1', url, dest_dir],
@@ -48,10 +49,9 @@ def download(url):
     """Downloads a module that is not a git repository"""
     parsed_url = urlparse(url)
     basename = os.path.basename(parsed_url.path)
-    extension = basename[-3:]
     response = urllib.request.urlopen(url)
     data = response.read()
-    filename = os.path.join(MOD_DIR, extension, basename)
+    filename = os.path.join(MOD_DIR, basename)
     if os.path.isfile(filename):
         raise Exception("File already exists: %s" % filename)
     with open(filename, 'wb') as f: